bug 1667213: move settings/versions to django page
This moves the settings/versions information that was on the front page
of the site to a Django admin page.

In order to do that, I had to add the ability to do static pages in the
Django admin.
