+++ tests/test_api.py
@@ -3,7 +3,6 @@
 # file, you can obtain one at http://mozilla.org/MPL/2.0/.
 
 import datetime
-import json
 
 import pytest
 
@@ -1024,63 +1023,6 @@ def test_stats_uploads(client):
     assert data["uploads"]["this_month"]["files"] == 100
     assert data["uploads"]["this_month"]["total_size"] == 123_456_789
     assert data["uploads"]["this_month"]["total_size_human"] == "117.7 MB"
-
-
-@pytest.mark.django_db
-def test_current_settings(client, settings):
-    settings.SYMBOL_URLS = [
-        "https://awsamazon.com/default-bucket-name",
-        "https://username:password@awsamazon.com/other-bucket-name",
-    ]
-    url = reverse("api:current_settings")
-    response = client.get(url)
-    assert response.status_code == 403
-
-    user = User.objects.create(username="peterbe", email="peterbe@example.com")
-    user.set_password("secret")
-    user.save()
-    assert client.login(username="peterbe", password="secret")
-
-    response = client.get(url)
-    assert response.status_code == 403
-
-    user.is_superuser = True
-    user.save()
-    response = client.get(url)
-    assert response.status_code == 200
-    current_settings = {x["key"]: x["value"] for x in response.json()["settings"]}
-    assert current_settings["SYMBOL_URLS"] == json.dumps(
-        [
-            "https://awsamazon.com/default-bucket-name",
-            "https://user:xxxxxx@awsamazon.com/other-bucket-name",
-        ]
-    )
-
-
-@pytest.mark.django_db
-def test_current_versions(client):
-    url = reverse("api:current_versions")
-    response = client.get(url)
-    assert response.status_code == 403
-
-    user = User.objects.create(username="peterbe", email="peterbe@example.com")
-    user.set_password("secret")
-    user.save()
-    assert client.login(username="peterbe", password="secret")
-
-    response = client.get(url)
-    assert response.status_code == 403
-
-    user.is_superuser = True
-    user.save()
-    response = client.get(url)
-    assert response.status_code == 200
-    current_versions = {x["key"]: x["value"] for x in response.json()["versions"]}
-    assert "Django" in current_versions
-    assert "Tecken" in current_versions
-    assert "PostgreSQL" in current_versions
-    assert "Redis Cache" in current_versions
-    assert "Redis Store" in current_versions
 
 
 @pytest.mark.django_db
